[#84] Apply cookie to episode's downloading process
- added new config for detecting source-type
- done small refactoring
