+++ src/modules/podcast/models.py
@@ -12,6 +12,7 @@ from core import settings
 from core.database import ModelBase
 from common.models import ModelMixin
 from common.db_utils import EnumTypeColumn
+from modules.podcast.utils import SourceType
 
 
 class EpisodeStatus(enum.Enum):
@@ -27,15 +28,6 @@ class EpisodeStatus(enum.Enum):
     DL_EPISODE_UPLOADING = "episode_uploading"
     DL_COVER_DOWNLOADING = "cover_downloading"
     DL_COVER_UPLOADING = "cover_uploading"
-
-    def __str__(self):
-        return self.value
-
-
-class SourceType(enum.Enum):
-    YOUTUBE = "YOUTUBE"
-    YANDEX = "YANDEX"
-    UPLOAD = "UPLOAD"
 
     def __str__(self):
         return self.value
