+++ src/tests/conftest.py
@@ -10,7 +10,8 @@ from sqlalchemy.ext.asyncio import AsyncSession
 
 from core import settings
 from modules.auth.models import UserInvite
-from modules.podcast.models import Podcast, Episode, Cookie, SourceType
+from modules.podcast.models import Podcast, Episode, Cookie
+from modules.podcast.utils import SourceType
 from modules.providers import utils as youtube_utils
 from tests.helpers import (
     PodcastTestClient,
